+++ MetarSharp/Methods/Convert/Pressure/FromInchesMercury.cs
@@ -14,6 +14,6 @@
         /// </summary>
         /// <param name="value"></param>
         /// <returns>the value converted to Hectopascals as a decimal with set decimal places</returns>
-        public static decimal ToHectopascals(this double value, byte decimalPlaces) => Math.Round((decimal)value * (decimal)33.86388666666671, decimalPlaces);
+        public static decimal ToHectopascals(this double value, int decimalPlaces) => Math.Round((decimal)value * (decimal)33.86388666666671, decimalPlaces);
     }
 }
