+++ MetarSharp/Methods/Convert/Time/FromWeeks.cs
@@ -22,18 +22,18 @@
 
         public static decimal ToHours(this double value) => Math.Round((decimal)value * 168, 2);
 
-        public static decimal ToHours(this double value, byte decimalPlaces) => Math.Round((decimal)value * 168, decimalPlaces);
+        public static decimal ToHours(this double value, int decimalPlaces) => Math.Round((decimal)value * 168, decimalPlaces);
 
         public static decimal ToDays(this double value) => Math.Round((decimal)value * 7, 2);
 
-        public static decimal ToDays(this double value, byte decimalPlaces) => Math.Round((decimal)value * 7, decimalPlaces);
+        public static decimal ToDays(this double value, int decimalPlaces) => Math.Round((decimal)value * 7, decimalPlaces);
 
         public static decimal ToMonths(this double value) => Math.Round((decimal)value / (decimal)4.345, 2);
 
-        public static decimal ToMonths(this double value, byte decimalPlaces) => Math.Round((decimal)value / (decimal)4.345, decimalPlaces);
+        public static decimal ToMonths(this double value, int decimalPlaces) => Math.Round((decimal)value / (decimal)4.345, decimalPlaces);
 
         public static decimal ToYears(this double value) => Math.Round((decimal)value / (decimal)52.143, 2);
 
-        public static decimal ToYears(this double value, byte decimalPlaces) => Math.Round((decimal)value / (decimal)52.143, decimalPlaces);
+        public static decimal ToYears(this double value, int decimalPlaces) => Math.Round((decimal)value / (decimal)52.143, decimalPlaces);
     }
 }
