+++ MetarSharp/Methods/Convert/Time/FromYears.cs
@@ -22,18 +22,18 @@
 
         public static decimal ToHours(this double value) => Math.Round((decimal)value * 8760, 2);
 
-        public static decimal ToHours(this double value, byte decimalPlaces) => Math.Round((decimal)value * 8760, decimalPlaces);
+        public static decimal ToHours(this double value, int decimalPlaces) => Math.Round((decimal)value * 8760, decimalPlaces);
 
         public static decimal ToDays(this double value) => Math.Round((decimal)value * 365, 2);
 
-        public static decimal ToDays(this double value, byte decimalPlaces) => Math.Round((decimal)value * 365, decimalPlaces);
+        public static decimal ToDays(this double value, int decimalPlaces) => Math.Round((decimal)value * 365, decimalPlaces);
 
         public static decimal ToWeeks(this double value) => Math.Round((decimal)value * (decimal)52.143, 2);
 
-        public static decimal ToWeeks(this double value, byte decimalPlaces) => Math.Round((decimal)value * (decimal)52.143, decimalPlaces);
+        public static decimal ToWeeks(this double value, int decimalPlaces) => Math.Round((decimal)value * (decimal)52.143, decimalPlaces);
 
         public static decimal ToMonths(this double value) => Math.Round((decimal)value * 12, 2);
 
-        public static decimal ToMonths(this double value, byte decimalPlaces) => Math.Round((decimal)value * 12, decimalPlaces);
+        public static decimal ToMonths(this double value, int decimalPlaces) => Math.Round((decimal)value * 12, decimalPlaces);
     }
 }
